sf: fix SurfaceView Dim layer show black issue
Gallery play video, press back key to end the playing, the sceen
flash black. SurfaceView Dim is the child of SurfaceView, but the
alphia do not change with the parent layer at the exiting. Thus is
the SurfaceView Dim layer do change the alpha along with the
transition animation. getAlpha can fix this issue, getAlpha with
check the parent alpha.

Change-Id: I090a410ce08d4054c8f67fec955173c13de1e2b6